bugfix: remove __versioned__ = {} on models
it was not possible to create a record with a file because of missmatching
version_id's. but the problem does not occure every time. sometimes it worked.
rdm-records does not have those statments, therefore this commit removes it and
the bug is gone!

in the end the bug is not really fixed but it works now!