feature: preview and download file
it was necessary to copy paste a lot of code from invenio-app-rdm

some minor changes were necessary to create the view. the biggest problem was to
reuse as many code as possible from invenio-app-rdm. But to do that it was
necessary to copy paste it. That should be fixed somehow in the future.

There were missing some routes to create a download link for marc21 records.