fix: the missing mappings causes an error with the empty search feature
The change is necessary to get rid of following error message:

Text fields are not optimised for operations that require per-document field
data like aggregations and sorting, so these operations are disabled by default.
Please use a keyword field instead. Alternatively, set fielddata=true on
[created] in order to load field data by uninverting the inverted index. Note
that this can use significant memory.

This fix is a quick fix. it should be refactored to be a real mapping.